Start using JS for a nice UX. While Crystal is incredibly fast, it's nice to not force users to reload pages and do navigation manually while interacting with the site. Most interactions dealing with races will thus be done asynchronously.
Should (read: when) things move to WebSocket communication, this will become a non-issue, but there will definitely still be JavaScript running on the client to deal with updates. In this scenario, all Race interactions will be done on a single page session, with all communication about the Race being done over WS. Things like actual registration (not guest logins) will still be done synchronously, for the sake of keeping simple things simple - as they should be.