Move to Toro/Crecto instead of Kemal. Kemal is great, but even it has more features than I really need, and typing out routes gets a little repetitive without native scoping support.
Toro makes having multiple routes under the same scope extremely easy, and even allows for partials of route handlers to be placed anywhere in the hierarchy. It still lacks some nice features like websocket handlers and having routes nested under a method block (`get` or `post`), but those are simple to work around.

Crecto is a Crystal port of Ecto that uses a similar approach to databases as opposed to the ActiveRecord pattern. It's one of the most feature-complete ORMs for crystal as of now, and works well in this use case.